backend/requirements.txt
asgiref==3.6.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
Django==4.1
django-anymail==11.1
django-cors-headers==3.13.0
django-model-utils==4.5.1
djangorestframework==3.13.1
Faker==26.0.0
gunicorn==20.1.0
idna==3.7
mysqlclient==2.1.1
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2022.7
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.4.3
stripe==10.6.0
typing_extensions==4.12.2
tzdata==2022.7
urllib3==2.2.2
